+++ SFRecordCompareEngine.Core/Database/Interfaces/IDatabaseSchemaInitializer.cs
@@ -2,5 +2,5 @@ namespace SFRecordCompareEngine.Core.Database.Interfaces;
 
 public interface IDatabaseSchemaInitializer
 {
-    void Initialize();
+    bool Initialize();
 }